[ALS-5583] Add Jenkins Docker UI upload job and improve pipeline
This commit adds a new Jenkins Docker job for uploading the pic-sure-ui Docker image to AWS S3. It reconfigures the biodatacatalyst-ui and PIC-SURE Pipeline jobs to integrate with the new upload job. It gets the tag of the Docker image from biodatacatalyst-ui build and passes it through the pipeline. The PIC-SURE Open Build job has been repurposed as an override build, which may change the Docker tag.