[ALS-5583] Reorder Jenkins pipeline stages
The stages 'PSAMA and PIC-SURE HPDS' and 'PIC-SURE Dictionary image' within the Jenkins pipeline configuration have been reordered. This was done to improve the logical flow of the pipeline - making it easier for users to understand the progression of each stage within the pipeline job. This refactor does not change the functionality of each stage.