Extract HTTPD instance private IP in Terraform script
The commit introduces extraction of private IP specifically for AWS HTTPD instances in the Terraform destroy standalone stack job. This change now allows for private IP logging of HTTPD instances in addition to the existing instances. It aims to provide more specific details about the infrastructure when running the destruction process.