Update default git hash and add conditional variables in flyway migration
The default git hash for Jenkins has been altered to point to a different branch, and a credential ID was added for GitLab token authentication in Jenkins Docker configuration. In the flyway migration script, conditions were added to include or exclude certain database components in the migration process based on the received flags.