Update error handling in flyway migration script (#111)
* Update error handling in flyway migration script

Added a status check after running the Flyway Docker command to catch any possible failures and to ensure successful migration. This was necessary after discovering that recent Flyway Docker image updates were missing the required MySQL drivers causing migration failures. Additionally, if IT does not configure the network appropriately we will now fail if we have a connection timeout.

* Refactor string encoding function in Bash_Functions job configuration

The string encoding function in the config file for the Bash_Functions job in jenkins-docker has been rewritten. Instead of a series of string substitutions, a for loop with a case statement is now used to handle different character scenarios including special symbols '?, #, {' for more comprehensive encoding.