[ALS-5387] Persist query ID code change: PIC-SURE will no longer store URL’s with stack information (#113)
* Update default git hash and add conditional variables in flyway migration

The default git hash for Jenkins has been altered to point to a different branch, and a credential ID was added for GitLab token authentication in Jenkins Docker configuration. In the flyway migration script, conditions were added to include or exclude certain database components in the migration process based on the received flags.

* Remove UUID assignment in Jenkins job configuration

The specific UUID assignment for target stacks 'a' and 'b' has been removed from the configuration script of the 'Teardown and Rebuild Stage Environment' Jenkins job. Subsequent Terraform commands no longer include var flag for 'dictionary_resource_id' and 'auth_hpds_resource_id', indicating that these may be assigned elsewhere or are not necessary anymore.