Add 'target_stack' parameter, update script references in Jenkins job
Added a new 'target_stack' parameter to the job configuration for deploying PSAMA to an existing Wildfly EC2. The script references, which were hardcoded for 'A stack', are now updated to refer to the 'target_stack' parameter. Also, moved the 'reset_role' function to the end of the script for a cleaner exit.