Add Jenkins jobs for downloading and uploading picsure-dictionary env (#135)
* Add Jenkins jobs for downloading and uploading picsure-dictionary env

Introduced two new Jenkins jobs: one for downloading the picsure-dictionary.env file from S3 and another for uploading it. These jobs automate the process of fetching and storing configuration files for the picsure-dictionary application.

* Update S3 upload script comments

Clarified file names in S3 upload script comments for both downloading and uploading picsure-dictionary configurations. This ensures that the comments accurately reflect the actual file handling in the scripts.