Als 8567 (#151)
* Update volume mount to include SELinux context option

The mount option ":Z" was added to ensure compatibility with SELinux-enabled systems by applying the correct context labeling for shared volumes. This change resolves potential permission issues when running the container on systems with SELinux policies enforced.

* Set threshold to NONE for ReverseBuildTrigger.

Updated the ReverseBuildTrigger threshold configuration to "NONE" with an ordinal of -1 and color GREY. This change ensures the trigger will not activate based on the upstream project’s success state.

---------

Co-authored-by: thomas_desain1 <thomas_desain@hms-srce.org>